@@ -4,16 +4,15 @@ description: >-
   Client code to start and interact with the frontend_server compiler from the
   Dart SDK.
 environment:
-  sdk: ">=2.10.0 <2.12.0"
+  sdk: ">=2.12.0-0 <2.14.0"
 dependencies:
-  async: ^2.0.0
-  meta: ^1.1.0
-  path: ^1.7.0
+  async: ^2.5.0
+  path: ^1.8.0
 dev_dependencies:
-  package_config: ^1.9.2
-  shelf: ^0.7.5
-  shelf_packages_handler: ^2.0.0
-  shelf_static: ^0.2.8
-  test: ^1.0.0
-  test_descriptor: ^1.0.0
-  vm_service: ^4.1.0
+  package_config: ^2.0.0
+  shelf: ^1.0.0
+  shelf_packages_handler: ^2.0.1
+  shelf_static: ^0.2.6
+  test: ^1.16.0
+  test_descriptor: ^2.0.0
+  vm_service: ^6.1.0
